requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.1
WTForms==3.1.2
Werkzeug==3.0.4
Jinja2==3.1.4
SQLAlchemy==2.0.32
Pillow==10.4.0
pyodbc==5.1.0
pymssql==2.2.5
